Improvements to app functionality
1. Improved flash messages display.
2. Improved route handling.
3. Improved admin route code.
4. flash messages jinja code block centralized to layout page.

+++ app/main/routes.py
@@ -1,6 +1,7 @@
-from flask import Blueprint, render_template, request, jsonify, session, flash, redirect, url_for
+from flask import Blueprint, render_template, request, jsonify, session, flash, redirect, url_for, send_from_directory
 from flask_login import current_user, login_required
 from app.extensions import db
+import os, textwrap
 
 from app.models import Book, Cart
 
@@ -59,16 +60,25 @@ def book_details(book_id):
 @login_required
 def add_to_cart(book_id):
     cart_item = Cart.query.filter_by(user_id = current_user.id, book_id=book_id).first()
+    book = Book.query.get(book_id)
+
+    if not book:
+        flash("Book not found", "danger")
+        return redirect(request.referrer)
+    
+    title = book.title
+    short_title = textwrap.shorten(title,width=30, placeholder="...." )
+    
     if cart_item:
         cart_item.quantity += 1
-        flash("Updated quantity for book", "info")
+        flash(f"Updated quantity for Book: '{short_title}' ", "info")
     else:
         new_item = Cart(user_id=current_user.id, book_id=book_id) #type: ignore
         db.session.add(new_item)
-        flash("Book added to cart!", "success")
-
+        flash(f" Book: '{short_title}' Added to cart", "success")
+    
     db.session.commit()
-    return redirect(request.referrer)
+    return redirect(request.referrer if request.referrer else url_for("main.home"))
 
 
 @main.route("/check_cart", methods=["GET", "POST"])
@@ -88,16 +98,32 @@ def check_cart():
 def delete_from_cart(book_id):
     
     cart_item = Cart.query.filter_by(user_id=current_user.id, book_id=book_id).first()
+    book = Book.query.get(book_id)
+
+    if not book:
+        flash("Book not found", "danger")
+        return redirect("/check_cart")
+    
+    title = book.title #type:ignore
+    short_title = textwrap.shorten(title,width=30, placeholder="...." )
+
     if cart_item:
         if cart_item.quantity > 1:
             cart_item.quantity -= 1
         else:
             db.session.delete(cart_item)
-            
+
+        flash(f" Book: '{short_title}' Deleted from cart", "info")
         db.session.commit()
-        flash("Book deleted from cart!", "info")
     else:
-        flash("Book not found in cart", "warning")
+        flash(f" Book: '{short_title}' not found in cart", "warning")
 
 
     return redirect("/check_cart")
+
+
+
+@main.route('/favicon.ico')
+def favicon():
+    return send_from_directory(os.path.join(main.root_path, 'static'),
+                               'favicon.ico', mimetype='image/vnd.microsoft.icon')
